requirements.txt
# Core AI and API dependencies
openai>=1.0.0
azure-openai

# Environment and configuration management
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing and utilities
requests>=2.31.0
json5>=0.9.0
python-dateutil>=2.8.0

# CLI interface enhancements
colorama>=0.4.6
rich>=13.0.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Logging and monitoring
structlog>=22.0.0
python-json-logger>=2.0.0

# Security and validation
cryptography>=3.4.8
validators>=0.20.0

# Optional: Web interface (future enhancement)
# flask>=2.3.0
# fastapi>=0.100.0
# uvicorn>=0.22.0

# Optional: Database integration (future enhancement)  
# sqlalchemy>=2.0.0
# sqlite3  # Built into Python

# Optional: Advanced NLP features (future enhancement)
# nltk>=3.8.0
# spacy>=3.6.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
